[IA64] Fix mca handler so as not to destroy ar
This patchf ixes the mca handler so as not to destroy ar
and some bugs.
On mca the firmware saves non-banked and banked general registers
and the most of ar registers must be unchanged.
However before ia64_os_mca_proc_state_dump saves such registers,
the cited hunks unconditionally overwrite ar.k[36] so that
they are destroyed.

Signed-off-by: Kazuhiro Suzuki <kaz@jp.fujitsu.com>